Refactor logging manager and update build
