+++ src/api/server.cpp
@@ -1,4 +1,4 @@
-#include "memory/manager.h"
+#include "core/logging.h"
 #include "memory/memory_tier_manager.hpp"
 #include "crow/crow_isolation.h"
 #include "api/crow_request.h"
