+++ src/memory/CMakeLists.txt
@@ -2,7 +2,6 @@ add_library(sep_memory
   STATIC 
     memory_tier.cpp
     memory_tier_manager.cpp
-    manager.cpp
     redis_manager.cpp
     quantum_coherence_manager.cpp
 )
