+++ src/quantum/pattern_processor_interface.cpp
@@ -7,7 +7,7 @@
 #include "core/common.h"  // defines sep::SEPResult
 #include "compat/core.h"
 #include "memory/memory_tier_manager.hpp"
-#include "memory/manager.h"  // This is actually the logging manager
+#include "core/logging.h"  // logging manager
 
 // Define namespace alias for clarity
 namespace logging = sep::logging;
